requirements.txt
appdirs==1.4.3
astroid==2.4.2
backcall==0.2.0
beautifulsoup4==4.9.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
joblib==1.0.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
matplotlib==3.3.3
mccabe==0.6.1
msgpack==0.6.2
nltk==3.5
numpy==1.19.5
packaging==20.3
parso==0.8.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
progress==1.5
prompt-toolkit==3.0.14
ptyprocess==0.7.0
Pygments==2.7.4
pylint==2.6.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pyzmq==22.0.2
regex==2020.11.13
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==2.1
toml==0.10.2
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
wordcloud==1.8.1
wrapt==1.12.1
